refactor: 출력 관련 함수 모듈화 - '구현 목록: 출력' 관련 함수 따로 분리 - 이름 입력 버튼을 누르면 결과 리셋
this refers #9, #7
