+++ src/index.js
@@ -1,9 +1,9 @@
 import RacingCarGame from './racingCarGame.js';
 import Scanner from './scanner.js';
+import Printer from './printer.js';
 
 const racingCarGame = new RacingCarGame();
 
-// TODO: 이름 다시 세팅하면 게임 결과 지우기
 const handleCarNamesSubmitBtn = function () {
   try {
     const carNamesList = Scanner.getCarNamesInputToList();
@@ -12,6 +12,7 @@ const handleCarNamesSubmitBtn = function () {
       document.querySelector('#racing-count-container'),
       true,
     );
+    Printer.initCarGameResultContainer();
   } catch (e) {
     alert(e.message);
     if (e.errElem) {
@@ -20,44 +21,23 @@ const handleCarNamesSubmitBtn = function () {
   }
 };
 
-const printRoundResult = function (printContainer) {
-  const carGameResultContainer = printContainer;
-  const oneRoundResultDiv = document.createElement('div');
-  racingCarGame.carList.forEach(car => {
-    oneRoundResultDiv.innerHTML += `
-    ${car.getName()}: ${'-'.repeat(car.getPosition())} <br/>`;
-  });
-  oneRoundResultDiv.innerHTML += '<br/>';
-  carGameResultContainer.appendChild(oneRoundResultDiv);
-};
-
-const printWinner = function (printContainer) {
-  const carGameResultContainer = printContainer;
-  const winnerResultDiv = document.createElement('div');
-  const winPosition = Math.max(
-    ...racingCarGame.carList.map(car => car.getPosition()),
-  );
-  const winners = racingCarGame.carList
-    .filter(car => car.getPosition() === winPosition)
-    .map(car => car.getName());
-  winnerResultDiv.innerHTML = `최종 우승자: ${winners.join(', ')}`;
-  carGameResultContainer.appendChild(winnerResultDiv);
-};
-
 const handleCountSubmitBtn = function () {
   try {
     const racingCount = Scanner.getRacingCountFromUser();
     const carGameResultContainer = document.querySelector(
       '#car-game-result-container',
     );
-    carGameResultContainer.innerHTML = '<h4>📄 실행 결과</h4>';
     racingCarGame.setRacingCount(racingCount);
     racingCarGame.clearCarPositions();
+    Printer.initCarGameResultContainer(carGameResultContainer);
     while (racingCarGame.racingCount--) {
       racingCarGame.raceOneRound();
-      printRoundResult(carGameResultContainer);
+      Printer.printOneRoundResult(
+        carGameResultContainer,
+        racingCarGame.carList,
+      );
     }
-    printWinner(carGameResultContainer);
+    Printer.printWinner(carGameResultContainer, racingCarGame.carList);
     Scanner.setElemVisible(carGameResultContainer, true);
   } catch (e) {
     // TODO: 알림 메소드 따로 빼기
